refactor(LdPlatform): extract to separate file
