+++ packages/liquid_flutter/lib/src/theme/liquid_material_theme.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:liquid_flutter/src/theme/platform.dart';
 import 'package:liquid_flutter/src/tokens.dart';
 import 'theme.dart';
